requirements.txt
torchvision
torch
pillow
numpy
einops
colorama
torchsummary
tqdm
tensorboard
pytest
config
decord
opencv-python
imageio